refactor(lint): split harn-lint/src/lib.rs into focused modules
lib.rs grew to 2787 lines mixing the stateful Linter walk with
comment scanning, source-aware rules, and small utility helpers.
Split it so lib.rs is just the public API (173 lines) and each
concern lives next to related code:

- diagnostic.rs: LintDiagnostic, LintSeverity, LintOptions
- decls.rs: per-walk tracking records
- naming.rs: case conversion + item classifiers
- harndoc.rs: comment scanning + legacy-doc-comment rule
- linter.rs + linter/walk.rs: Linter struct, helpers, finalize,
  and the large lint_node match
- rules/: one file per source-aware rule family

Public API is preserved (all harn_lint::lint_* entry points,
LintDiagnostic, LintSeverity, LintOptions, DEFAULT_COMPLEXITY_THRESHOLD,
derive_file_header_title, simplify_bool_comparison,
collect_selective_import_names). Closes #43.
